Add files via upload
For a recent boot camp project, we selected data from Data.gov to address a real-world problem. Our project involved creating a dashboard in Excel, which unfortunately is too large to list here. We utilized Power Query and 3D maps to analyse the data. Our goal was to identify the most frequent area in Los Angeles for road traffic accidents over a significant period.